fix: eslint fixes
